Create Java console application to connect API info
This code uses the java.net.URL and java.net.HttpURLConnection classes to connect to the Binance API and retrieve the current price of BTC in USDT. The code sends a GET request to the specified URL and then prints the response code and the response from the API.

You can use the similar method to check the price change and use similar methods as the python code snippet provided earlier.

Make sure you have the java.io package imported, and also make sure that the Java Development Kit (JDK) is installed on your system, so you can compile and run the application.

You can run the above code by using the command javac BinanceAPI.java to compile the code and java BinanceAPI to run the application.

Please note that this is just a basic example, and you may need to adjust the code to suit your specific needs. Also, make sure that you are using the correct endpoint for retrieving historical data and that your API key is valid.